Project Skeleton with RESTful Service
Project containing the skeleton for each layer. The presentation layer
uses jersey servlets to generate JSON. A stub data access layer has been
created. This stub only implement one method of the IDAL for now
allowing to get all the events. This implementation only creates a few
events and return the list to the upper layer: the business layer